Updated error handling and snackbar messages for all(?) rpcs and vue components

+++ grpc/src/main/java/edu/cooper/ece366/restaurantReservation/grpc/Users/UserManager.java
@@ -75,7 +75,7 @@ public class UserManager {
             return dao.getIdByUsername(username);
         });
         if (userId.isPresent())
-            return userId.get().intValue();
+            return userId.get();
         else
             throw new InvalidUsernameException("Username does not exist.");
     }
